mqtt-client/requirements.txt
influxdb-client
python-dotenv
paho-mqtt
pytz
